Extract RailGui
